set_start_method("fork") for mac OS with python 3.8+
