+++ tests/test_cli.py
@@ -4,6 +4,7 @@ import os
 
 from click.testing import CliRunner
 from mlchain.cli.main import main
+from .utils import test_breaking_process
 
 logger = logging.getLogger()
 
@@ -35,37 +36,4 @@ class TestCLI(unittest.TestCase):
             return 0
         cli = main(is_testing=True)
         runner = CliRunner()
-        test_breaking_process(runner, cli, args='run'.split(), new_pwd='tests/dummy_server', prog_name='python -m mlchain')
-
-
-def test_breaking_process(runner, cli, args, new_pwd, prog_name):
-    from multiprocessing import Queue, Process
-    from threading import Timer
-    from time import sleep
-    from os import kill, getpid
-    from signal import SIGINT
-
-    q = Queue()
-
-    # Running out app in SubProcess and after a while using signal sending 
-    # SIGINT, results passed back via channel/queue  
-    def background():
-        Timer(5, lambda: kill(getpid(), SIGINT)).start()
-        os.chdir(new_pwd)
-        result = runner.invoke(cli, args, prog_name=prog_name)
-        q.put(('exit_code', result.exit_code))
-        q.put(('output', result.output))
-
-    p = Process(target=background)
-    p.start()
-
-    results = {}
-
-    while p.is_alive():
-        sleep(0.5)
-    else:
-        while not q.empty():
-            key, value = q.get()
-            results[key] = value
-    logger.info('Output of `mlchain run`:\n' + results['output'])
-    assert results['exit_code'] == 0
+        test_breaking_process(runner, cli, args='run'.split(), new_pwd='tests/dummy_server', prog_name='python3 -m mlchain')
